api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "author-kit-api"
version = "0.1.0"
description = "AuthorKit HTTP API — creative writing assistant backend"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Julien Moulin", email = "julien@supralab.fr" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.2.0",
    "langchain-together>=0.2.0",
    "requests>=2.31.0",
    "tiktoken>=0.7.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["author_kit"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
